Update RandomNumGen.java
fixed overflow issue causing negative values

+++ RandomNumGen.java
@@ -8,7 +8,7 @@ public class RandomNumGen {
     private static final int increment = 3517; // increment c
     private static final int modulus = (int) Math.pow(2, 17); //modulus K
 
-    private int current; //current value in the sequence
+    private long current; //current value in the sequence
 
     public RandomNumGen() {
         this.current = seed; // initializing w/ the seed value
